reprts

+++ Models/ModelsView/DistributorInTransaction.cs
@@ -8,7 +8,7 @@ namespace GasStation.Models.ModelsView
     public class DistributorInTransaction : Distributor
     {
         public string NameOfFuel { get; set; }
-        public float PriceForLiter { get; set; }
+        public decimal PriceForLiter { get; set; }
         public bool InTransaction { get; set; }
         public float Sum { get; set; }
         public int TankId { get; set; }
